+++ app-server/src/main/java/com/growth/task/mypage/controller/MyPageController.java
@@ -1,5 +1,8 @@
 package com.growth.task.mypage.controller;
 
+import com.growth.task.review.dto.ReviewStatsRequest;
+import com.growth.task.review.dto.ReviewStatsResponse;
+import com.growth.task.review.service.ReviewListService;
 import com.growth.task.todo.application.TodoListService;
 import com.growth.task.todo.dto.TodoListRequest;
 import com.growth.task.todo.dto.TodoStatsRequest;
@@ -20,16 +23,22 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/v1/mypage")
+@RequestMapping("/api/v1/mypage/{user_id}")
 @Tag(name = "MyPage", description = "My Page API Document")
 public class MyPageController {
     private final TodoListService todoListService;
+    private final ReviewListService reviewListService;
 
-    public MyPageController(TodoListService todoListService) {
+    public MyPageController(
+            TodoListService todoListService,
+            ReviewListService reviewListService
+    ) {
         this.todoListService = todoListService;
+        this.reviewListService = reviewListService;
     }
+
     @Operation(summary = "사용자 Todo 통계")
-    @GetMapping("/{user_id}/todos/stats")
+    @GetMapping("/todos/stats")
     @ResponseStatus(HttpStatus.OK)
     public TodoStatsResponse getTodoCounts(
             @PathVariable("user_id") Long userId,
@@ -37,8 +46,9 @@ public class MyPageController {
     ) {
         return todoListService.getTodoStats(userId, request);
     }
+
     @Operation(summary = "사용자 Todo 통계 세부 리스트")
-    @GetMapping("/{user_id}/todos")
+    @GetMapping("/todos")
     @ResponseStatus(HttpStatus.OK)
     public Page<TodoDetailResponse> getTodosByUser(
             @PathVariable("user_id") Long userId,
@@ -46,5 +56,13 @@ public class MyPageController {
             @ModelAttribute TodoListRequest request
     ) {
         return todoListService.getTodoByUserAndParams(pageable, userId, request);
+    }
+    @GetMapping("/review/stats")
+    @ResponseStatus(HttpStatus.OK)
+    public ReviewStatsResponse getReviewStats(
+            @PathVariable("user_id") Long userId,
+            @ModelAttribute @Valid ReviewStatsRequest request
+    ){
+        return reviewListService.getReviewStats(userId, request);
     }
 }
